move the environment variables creation from the lambda construct to the api stack since they're all the same
